Fix removing file which use last cluster of 4GBytes
The FAT file system limit the file size to 4GBytes.
Physical size is how much space in storage the file
use, sometimes, we create a file which size is 10 Bytes.
But it use  4KB space in disk.
fsck_msdos use "int" type to calculate physical size of
a file. If a file use last cluster of 4GBytes, 4GB - 1 etc.
Then the physical size will overflow, and fsck_msdos treate
the size as 0 byte.

Change-Id: Ie76f2b8e0cbb510fa9d33c6a8543e6cded6bd347
Signed-off-by: Hriday Sharma <hridaysharma42@gmail.com>